+++ frontend/src/components/Home/Interventions.js
@@ -1,8 +1,8 @@
 // renders the Interventions dropdown menu in the left panel of Home view
 import React, { useState } from 'react';
-import Antivirals from './Antivirals';
-import Vaccine from './Vaccine';
-import NPIForm from './NPI';
+import AntiviralsForm from './AntiviralsForm';
+import VaccineForm from './VaccineForm';
+import NPIForm from './NPIForm';
 import { createPortal } from 'react-dom';
 import './SetScenario.css';
 
@@ -51,7 +51,7 @@ const Interventions = ( {counties, npiChange} ) => {
                 <span className="modal-close" onClick={closeAntivirals}>&times;</span>
                 <h2>Antiviral Parameters and Stockpiles</h2>
                 <br></br>
-                <Antivirals onSubmit={(eff, wf, avs) => { console.log("Eff: ", eff, " WF: ", wf, " AVS: ", avs); closeAntivirals(); }} />
+                <AntiviralsForm onSubmit={(eff, wf, avs) => { console.log("Eff: ", eff, " WF: ", wf, " AVS: ", avs); closeAntivirals(); }} />
               </div>
             </div>,
             document.body
@@ -65,7 +65,7 @@ const Interventions = ( {counties, npiChange} ) => {
               <div className="modal-contents" onClick={(e) => e.stopPropagation()}>
                 <span className="modal-close" onClick={closeVaccine}>&times;</span>
                 <h2>Vaccine Stockpile</h2>
-                <Vaccine onSubmit={(ve, va, vwf, vs, vsl) => { console.log("Eff:", ve, " Adh: ", va, " WF: ", vwf, " Strat: ", vs, " VS: ", vsl); closeVaccine(); }} />
+                <VaccineForm onSubmit={(ve, va, vwf, vs, vsl) => { console.log("Eff:", ve, " Adh: ", va, " WF: ", vwf, " Strat: ", vs, " VS: ", vsl); closeVaccine(); }} />
               </div>
             </div>,
             document.body
